refactor(routers): refactor user's router
I misunderstand the way `response_model` works in routers decorators so
each endpoint return the same type as the one defined in `response_model`.
So I had to convert each ORM result in FastAPI's schema.

This is why this refactor endpoints to make them return an ORM model instead
of a FastAPI's schema.